finish the function of updating groomer

+++ appointment_system/src/main/java/com/cpt202/appointment_system/Models/Groomer.java
@@ -6,6 +6,8 @@ import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
 
+import com.mysql.cj.protocol.a.NativeConstants.IntegerDataType;
+
 import lombok.Data;
 
 
@@ -33,7 +35,7 @@ public class Groomer {
     // since rank() is a buil-in function in MySQL 
     // need
     @Column(name = "ranking", columnDefinition = "tinyint", nullable = false)
-    private byte rank;
+    private Integer rank;
 
     @Column(columnDefinition = "varchar(11)", nullable = false)
     private String phoneNumber;
@@ -47,7 +49,7 @@ public class Groomer {
     }
 
     
-    public Groomer(int gid, String name, String gender, String imageURL, byte rank, String phoneNumber) {
+    public Groomer(int gid, String name, String gender, String imageURL, Integer rank, String phoneNumber) {
         this.gid = gid;
         this.name = name;
         this.gender = gender;
